Pull class into its own file
