+++ Unit4/Unit4/BcrReport.cs
@@ -12,7 +12,7 @@ namespace Unit4
 {
     internal class BcrReport
     {
-        private enum Tier { Tier3, Tier4, CostCentre };
+        public enum Tier { Tier3, Tier4, CostCentre };
 
         private readonly BCRLineBuilder _builder = new BCRLineBuilder();
         private readonly ILogging _log;
@@ -64,43 +64,6 @@ namespace Unit4
             }
 
             return bag;
-        }
-
-        private class Report
-        {
-            public Tier Tier { get; set; }
-            public IGrouping<string, CostCentre> Hierarchy { get; set; }
-
-            public string Parameter { get { return Hierarchy.Key; } }
-
-            public bool ShouldFallBack { get { return (Tier == Tier.Tier3 || Tier == Tier.Tier4) && Hierarchy.Any(); } }
-
-            public IEnumerable<Report> FallbackReports()
-            {
-                var fallbackGroups = Hierarchy.GroupBy(FallBackGroupingFunction(Tier), x => x);
-
-                return fallbackGroups.Select(x => new Report() { Tier = FallBackTier(Tier), Hierarchy = x });
-            }
-
-            private Tier FallBackTier(Tier current)
-            {
-                switch (current)
-                {
-                    case Tier.Tier3: return Tier.Tier4;
-                    case Tier.Tier4: return Tier.CostCentre;
-                    default: throw new InvalidOperationException("Should not fall back");
-                }
-            }
-
-            private Func<CostCentre, string> FallBackGroupingFunction(Tier current)
-            {
-                switch (current)
-                {
-                    case Tier.Tier3: return x => x.Tier4;
-                    case Tier.Tier4: return x => x.Code;
-                    default: throw new InvalidOperationException("Should not fall back");
-                }
-            }
         }
 
         private IEnumerable<BCRLine> RunBCR(Report report)
